fix: 메세지 한 번에 담아서 출력되게 수정
